Create Buttons.tsx component and summarised all buttons here
